# libs/contracts/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: DocVault API Gateway
  description: |
    Unified API gateway for DocVault — secure document management system.

    ## Authentication

    **Development (localhost):** After signing in via `/auth/login`, the gateway sets an
    `dv_access_token` cookie. Subsequent API calls automatically include this cookie.

    **Production:** Pass the Keycloak JWT in the `Authorization` header:
    `Authorization: Bearer <token>`

    Both methods are accepted on all protected endpoints.

    ## Roles
    | Role | Permissions |
    |------|-------------|
    | `viewer` | Read published documents, download files |
    | `editor` | Create, upload, submit, archive documents |
    | `approver` | Approve/reject pending documents |
    | `compliance_officer` | Read audit logs and allowed metadata; cannot preview, stream, presign, or download file content |
    | `admin` | Full access |

    ## Document Lifecycle
    ```
    DRAFT → PENDING → PUBLISHED → ARCHIVED
    ```

    ## Compliance Officer Restriction
    Compliance officers (`compliance_officer`) are **never allowed** to preview,
    stream, presign, or download file content, regardless of ACL configuration —
    enforced by route guards and policy.service.ts.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.docvault.internal
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Keycloak-issued RS256 JWT token passed as `Authorization: Bearer <token>`"
    cookieAuth:
      type: apiKey
      in: cookie
      name: dv_access_token
      description: JWT issued by the gateway after Keycloak OIDC login — set automatically by the frontend
  schemas:
...
    NotificationRecord:
      type: object
      required: [id, type, docId, recipientId, createdAt, read]
      properties:
        id:          { type: string, description: 'Unique notification ID' }
        type:        { type: string, enum: [SUBMITTED, APPROVED, REJECTED, ARCHIVED, DELETED] }
        docId:       { type: string }
        recipientId: { type: string, description: 'User ID this notification is addressed to' }
        docTitle:    { type: string, nullable: true, description: 'Human-readable document title for UI display' }
        reason:      { type: string, nullable: true }
        traceId:     { type: string, nullable: true }
        createdAt:   { type: string, format: date-time }
        read:        { type: boolean }
    Notification:
      type: object
      required: [type, docId]
      properties:
        type:
          type: string
          enum: [SUBMITTED, APPROVED, REJECTED, ARCHIVED, DELETED]
        docId:       { type: string }
        recipientId:
          type: string
          nullable: true
          description: 'Single recipient (mutually exclusive with recipientIds)'
        recipientIds:
          type: array
          items: { type: string }
          nullable: true
          description: 'Multiple recipients for SUBMITTED type'
        docTitle:    { type: string, nullable: true }
        reason:      { type: string, nullable: true }
    NotificationPage:
      type: object
      properties:
        records: { type: array, items: { $ref: '#/components/schemas/NotificationRecord' } }
        total:   { type: integer }
        page:    { type: integer }
        pages:   { type: integer }
    UnreadCount:
      type: object
      properties:
        count: { type: integer }

    # ─── Auth ─────────────────────────────────────────────────────────────────
...
paths:
...
  /notify:
    get:
      summary: List paginated notifications for the current user
      operationId: listNotifications
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Notifications]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
          description: 1-based page index
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
          description: Page size
      responses:
        '200':
          description: Paginated notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPage'
        '401':
          description: Unauthorized
    post:
      summary: Send a notification (internal — called by workflow-service)
      operationId: sendNotification
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Notification'
      responses:
        '200':
          description: Notification accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:   { type: boolean }
                  type:       { type: string }
                  docId:      { type: string }
                  recipients: { type: array, items: { type: string } }
        '401':
          description: Unauthorized
  /notify/mark-read:
    post:
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Notifications]
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        '401':
          description: Unauthorized
  /notify/unread-count:
    get:
      summary: Get unread notification count (lightweight — used for polling)
      operationId: getUnreadNotificationCount
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Notifications]
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnreadCount'
        '401':
          description: Unauthorized
  /notify/{id}/read:
    post:
      summary: Mark a single notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
        - cookieAuth: []
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Notification ID
      responses:
        '200':
          description: Done
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        '401':
          description: Unauthorized

# ─── Reusable Parameters ─────────────────────────────────────────────────────
...